// apps/client/src/app/arena/components/CommandConsole/ScriptEditor/types.ts
export interface Suggestion {
    label: string;
    detail: 'command' | 'control' | 'identifier' | 'flag' | string;
    hint: string;
}

export interface CaretPosition {
    bottom: number;
    left: number;
}

export interface ScriptEditorProps {
    scriptInput: string;
    setScriptInput: (val: string) => void;
    handleDeployBrain: () => void;
    toggleLibrary: () => void;
    clearPrebuilt: () => void;
}